# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartSpend API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Expense {
  id: string;
  title: string;
  amount: number;
  category: string;
  date: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface ExpenseInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 0 */
  amount: number;
  category: string;
  date: string;
  description?: string;
}

export interface ExpenseUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minimum 0 */
  amount?: number;
  category?: string;
  date?: string;
  description?: string;
}

export interface ExpenseListResponse {
  expenses: Expense[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CategoryLimit {
  category: string;
  limit: number;
  spent: number;
}

export interface Budget {
  id: string;
  month: number;
  year: number;
  totalLimit: number;
  categoryLimits: CategoryLimit[];
  spent: number;
  remaining: number;
}

export interface CategoryLimitInput {
  category: string;
  limit: number;
}

export interface BudgetInput {
  /** @minimum 0 */
  totalLimit: number;
  month?: number;
  year?: number;
  categoryLimits?: CategoryLimitInput[];
}

export interface DashboardSummary {
  totalSpentThisMonth: number;
  totalBudget: number;
  budgetUsedPercent: number;
  remainingBudget: number;
  expenseCount: number;
  /** @nullable */
  topCategory: string | null;
  recentExpenses: Expense[];
}

export interface MonthlySpend {
  month: string;
  total: number;
}

export interface CategorySpend {
  category: string;
  total: number;
  percent: number;
}

export type GetExpensesParams = {
category?: string;
startDate?: string;
endDate?: string;
page?: number;
limit?: number;
};