# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartSpend API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryLimitInput } from './categoryLimitInput';

export interface BudgetInput {
  /** @minimum 0 */
  totalLimit: number;
  month?: number;
  year?: number;
  categoryLimits?: CategoryLimitInput[];
}