# lib/api-zod/src/generated/types/categoryLimitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartSpend API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryLimitInput {
  category: string;
  limit: number;
}